+++ internal/db/connection.go
@@ -25,7 +25,13 @@ func Connect(uri string) (*mongo.Client, error) {
 
 	var result bson.M
 
-	if err := client.Database("admin").RunCommand(context.TODO(), bson.D{{"ping", 1}}).Decode(&result); err != nil {
+	if err := client.Database("admin").RunCommand(
+		context.TODO(),
+		bson.D{{
+			"ping",
+			1,
+		}}).
+		Decode(&result); err != nil {
 		return nil, err
 	}