+++ ui/template/templates.go
@@ -11,6 +11,7 @@ type Pages struct {
 	Base      func(title string, page templ.Component, data *shared.TemplateData) templ.Component
 	NotFound  func(data *shared.TemplateData) templ.Component
 	CheckList func(data *shared.TemplateData) templ.Component
+	UserLogin func(data *shared.TemplateData) templ.Component
 	// LogIn     func(data *shared.TemplateData) templ.Component
 }
 
@@ -24,6 +25,7 @@ func CreatePages() *Pages {
 		Base:      Base,
 		NotFound:  pages.NotFound,
 		CheckList: pages.CheckList,
+		UserLogin: pages.UserLogin,
 	}
 }